Let util.PathManager manage temporary directories
Add make_tempdir(), rm_tempdir() and cleanup() methods to
util.PathManager, wrapping tempfile's mkdtemp(). We store a list of
temporary directories in the PathManager Singleton so we can delete them
on program exit.
Register this cleanup() method in an abortHandler() method for the
default DataManager, which gets called at exit or interrupt.