Honest interval math in datelabel
Revamp the datelabel code, keeping the same interface. Date()s are now
defined as DateRange()s, instead of the other way around, with the
duration of the DateRange() being given by the Date's precision. Start
and end points are stored for DateRanges, and DateRanges have a single
precision attribute. DateRanges continue to be defined as [closed, open)
intervals.