Move defaults handling into CLIHandler
Add new partial_defaults argument to CLIHandler init. If set, this will
be a dict or list of dicts of settings that are given intermediate
priority between the explicitly set CLI options and their default
values (logic that was previously in FrameworkCLIHandler).

Split up parsing into preparse_cli(), which ignores the
partial_defaults, and parse_cli(), which takes them into account. This
is needed because in FrameworkCLIHandler, the source of partial_defaults
is the user input file which is passed on the command line, so we need
to do the preparsing to figure out where that is.