Remove convective_transition-specific code in shared_diagnostic
The legacy convective_transition POD set a
BULK_TROPOSPHERIC_TEMPERATURE_MEASURE env var in its settings.jsonc to
either '1' or '2' in order to select which input variable was used, but
the POD's code generated plots with that variable's name in the
filename. shared_diagnostic.make_pod_html had code to handle this
specific case, but that's hacky.

Implemented solution was to 1) change the env var defined in
settings.jsonc to refer to the variable name, 2) define the old env var
in convective_transition_diag_v1r3.py in terms of the new one, so the
remainder of the POD's code is unchanged, 3) update the POD's html
template to use the new env var.